requirements.txt
django==2.2.6
django-debug-toolbar==1.10.1
python-decouple==3.1
pytz==2018.5
sqlparse==0.2.4
django-imagekit==4.0.2
pillow==7.0.0
django-ckeditor==5.8.0
psycopg2==2.8.4
django-allauth